Added graph main page + tabs for three rooms
