+++ IntelliHouse2000App/Views/Climate/ClimateGraphsPage.xaml
@@ -1,92 +1,19 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<Shell xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              x:Class="IntelliHouse2000App.Views.ClimateGraphsPage"
-             xmlns:chart="clr-namespace:Syncfusion.Maui.Charts;assembly=Syncfusion.Maui.Charts"
-             xmlns:climate="clr-namespace:IntelliHouse2000App.ViewModels"
-             x:DataType="climate:ClimateGraphsPageViewModel">
-    <StackLayout>
-            <chart:SfCartesianChart BackgroundColor="LightGrey"
-                                    Title="Kitchen"
-                                    HeightRequest="250">
-                
-                <chart:SfCartesianChart.XAxes>
-                    <chart:CategoryAxis>
-                    </chart:CategoryAxis>
-                </chart:SfCartesianChart.XAxes>
-                <chart:SfCartesianChart.YAxes>
-                    <chart:NumericalAxis Minimum="0" Maximum="80" />
-                </chart:SfCartesianChart.YAxes>  
-
-                <chart:LineSeries ItemsSource="{Binding KitchenValues}"
-                                  XBindingPath="Timestamp" 
-                                  YBindingPath="Temperature"
-                                  EnableAnimation="True"
-                                  Fill="Red"
-                                  Label="Temperature"/>
-                <chart:LineSeries ItemsSource="{Binding KitchenValues}" 
-                                  XBindingPath="Timestamp"
-                                  YBindingPath="Humidity"
-                                  EnableAnimation="True"
-                                  Fill="Blue"
-                                  Label="Humidity"/>
-            </chart:SfCartesianChart>
-            
-            <chart:SfCartesianChart BackgroundColor="LightGrey"
-                                    Title="Bedroom"
-                                    HeightRequest="250">
-                
-                <chart:SfCartesianChart.XAxes>
-                    <chart:CategoryAxis >
-                    </chart:CategoryAxis>
-                </chart:SfCartesianChart.XAxes>
-                <chart:SfCartesianChart.YAxes>
-                    <chart:NumericalAxis Minimum="0" Maximum="80" />
-                </chart:SfCartesianChart.YAxes>  
-
-                <chart:LineSeries ItemsSource="{Binding BedroomValues}"
-                                  XBindingPath="Timestamp" 
-                                  YBindingPath="Temperature"
-                                  EnableAnimation="True"
-                                  Fill="Red"
-                                  Label="Temperature"/>
-                <chart:LineSeries ItemsSource="{Binding BedroomValues}" 
-                                  XBindingPath="Timestamp"
-                                  YBindingPath="Humidity"
-                                  EnableAnimation="True"
-                                  Fill="Blue"
-                                  Label="Humidity"/>
-            </chart:SfCartesianChart>
-            
-            <chart:SfCartesianChart BackgroundColor="LightGrey"
-                                    Title="Livingroom"
-                                    HeightRequest="280">
-                
-                <chart:SfCartesianChart.Legend>
-                    <chart:ChartLegend Placement="Bottom">
-                    </chart:ChartLegend>
-                </chart:SfCartesianChart.Legend>
-                
-                <chart:SfCartesianChart.XAxes>
-                    <chart:CategoryAxis >
-                    </chart:CategoryAxis>
-                </chart:SfCartesianChart.XAxes>
-                <chart:SfCartesianChart.YAxes>
-                    <chart:NumericalAxis Minimum="0" Maximum="80" />
-                </chart:SfCartesianChart.YAxes>   
-
-                <chart:LineSeries ItemsSource="{Binding LivingroomValues}"
-                                  XBindingPath="Timestamp" 
-                                  YBindingPath="Temperature"
-                                  EnableAnimation="True"
-                                  Fill="Red"
-                                  Label="Temperature"/>
-                <chart:LineSeries ItemsSource="{Binding LivingroomValues}" 
-                                  XBindingPath="Timestamp"
-                                  YBindingPath="Humidity"
-                                  EnableAnimation="True"
-                                  Fill="Blue"
-                                  Label="Humidity"/>
-            </chart:SfCartesianChart>
-    </StackLayout>
-</ContentPage>
+             xmlns:views="clr-namespace:IntelliHouse2000App.Views">
+    <TabBar>
+        <Tab Title="Logs">
+            <ShellContent 
+                Title="Kitchen"
+                ContentTemplate="{DataTemplate views:KitchenGraphsPage}" />
+            <ShellContent 
+                Title="Bedroom"
+                ContentTemplate="{DataTemplate views:BedroomGraphsPage}" />
+            <ShellContent 
+                Title="Livingroom"
+                ContentTemplate="{DataTemplate views:LivingroomGraphsPage}" />
+        </Tab>
+    </TabBar>
+</Shell>
