+++ IntelliHouse2000App/Views/Climate/ClimateGraphsPage.xaml.cs
@@ -1,24 +1,11 @@
-﻿using IntelliHouse2000App.ViewModels;
-
-namespace IntelliHouse2000App.Views;
+﻿namespace IntelliHouse2000App.Views;
 
 
-public partial class ClimateGraphsPage : ContentPage
+public partial class ClimateGraphsPage : Shell
 {
-	private ClimateGraphsPageViewModel _viewModel;
-	public ClimateGraphsPage(ClimateGraphsPageViewModel viewModel)
+	public ClimateGraphsPage()
 	{
-		_viewModel = viewModel;
-		BindingContext = _viewModel;
 		InitializeComponent();
-	}
-
-	protected override async void OnAppearing()
-	{
-		base.OnAppearing();
-		await _viewModel.GetBedroomAsync(DateTime.Now);
-		await _viewModel.GetKitchenAsync(DateTime.Now);
-		await _viewModel.GetLivingroomAsync(DateTime.Now);
 	}
 }
