+++ IntelliHouse2000App/Views/Home/LogMainPage.xaml.cs
@@ -1,5 +1,3 @@
-using IntelliHouse2000App.ViewModels.Home;
-
 namespace IntelliHouse2000App.Views;
 
 public partial class LogMainPage : Shell
